Split Payment View Model into Payment and PaymentRquest View Models
