+++ src/main/kotlin/exceptions/runtime/RestrictedMethodCallException.kt
@@ -1,0 +1,3 @@
+package exceptions.runtime
+
+class RestrictedMethodCallException : RuntimeException()
